[Gold II] Title: 피보나치 수 3, Time: 16 ms, Memory: 13740 KB -BaekjoonHub

+++ 백준/Gold/2749. 피보나치 수 3/피보나치 수 3.cc
@@ -4,12 +4,12 @@ typedef long long ll;
 const long long int MAX = 1000000;
 ll arr[1500000];
 int main(){
-    long long int n; 
+    ll n; 
     cin>>n;
 
     arr[0] = 0;
     arr[1] = 1;
-    for(long long int i=2; i<1500000; i++){
+    for(ll i=2; i<1500000; i++){
         arr[i] = (arr[i-1] + arr[i-2] ) % MAX;
     }
     cout<<arr[n%1500000];
